requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pytest-html>=4.0.0

# Code Analysis
memory-profiler>=0.61.0
radon>=6.0.0

# Fairness
aif360>=0.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0